pyproject.toml
[project]
name = "lab-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "click>=8.3.1",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pylint>=4.0.3",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
